Provide a new send_stop method for InstandShutdownServer
I'm not convinced that the server wouldn't have fired off a request in
a thread, checked the __shutdown_request and gone back to sleep before
the call to shutdown().

This way we have to set the variable directly, but it seems much safer.

As a side effect, the SIGHUP handler will no longer block waiting for
the web server to shut down.