Make exporter metrics global
It doesn't make sense to keep them on the Collector, since Prometheus
metrics are registered with the global registry.

As a side effect, any collector that goes unhealthy will cause other
collectors to go unhealthy. This exposes the flaw in reading the
counters to determine if we're still healthy or not.